Fix problem when requesting old tweets
Because I am filtering the displayed tweets, I can't simply
use the :last tweet's ID to fetch older tweets (well I could, but
it is inefficient). Hence I store the oldest fetched tweet ID
(regardless whether displayed or not), and request tweets older than
that.

There was a problem when a new tweet was loaded (newer than the
newest), which overwrote the last-tweet-id. This is now fixed